! Postgresql and sqlite don't do tinyint or unsigned (install/upgrade sql files) [Bug 3772]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@9061 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ other/install_2-0_postgresql.sql
@@ -1643,7 +1643,7 @@ CREATE TABLE {$db_prefix}members (
   ignore_boards text NOT NULL,
   warning smallint NOT NULL default '0',
   passwd_flood varchar(12) NOT NULL default '',
-  pm_receive_from tinyint(4) unsigned NOT NULL default '0',
+  pm_receive_from smallint NOT NULL default '0',
   PRIMARY KEY (id_member)
 );
