+++ other/install_2-0_sqlite.sql
@@ -1373,7 +1373,7 @@ CREATE TABLE {$db_prefix}members (
   ignore_boards text NOT NULL,
   warning smallint NOT NULL default '0',
   passwd_flood varchar(12) NOT NULL default '',
-  pm_receive_from tinyint(4) unsigned NOT NULL default '0'
+  pm_receive_from smallint NOT NULL default '0'
 );
 
 #
